+++ tests/Pictograms.svelte
